azd ai agents: remove broken doctor model-deployments check
The `remote.model-deployments` check compared `agent.manifest.yaml`'s
`resources[].name` against Foundry deployment names, but that `name`
is a manifest-local alias / token (used to bind `{{token}}` placeholders
in `template.environment_variables`), not a deployment name. On every
sample created via `azd ai agent init`, the check therefore reports
the literal alias (e.g. AZURE_AI_MODEL_DEPLOYMENT_NAME) as a "missing
deployment", which is incorrect and noisy.

The actual deployment name lives in:
  - `agent.yaml`'s resolved env var (e.g.
    `AZURE_AI_MODEL_DEPLOYMENT_NAME: gpt-4.1-mini`), and / or
  - the azd environment under the same key.

A redesign that reads from the right source will land in a follow-up.
For now, removing the false-failure check is the smallest correct
change.

Removed:
  - internal/cmd/doctor/checks_model_deployments.go (check + helpers
    newCheckModelDeployments, classifyModelDeployments,
    realProbeModelDeployments, parseAccountFromProjectID,
    modelDeploymentProbeFn, listDeploymentNames,
    modelDeploymentsProbeTimeout)
  - internal/cmd/doctor/checks_model_deployments_test.go

Updated:
  - checks_remote.go: drop registration, add tombstone comment
    pointing at redesign requirements for the next check
  - checks_local.go: drop unused probeModelDeployments seam from
    Dependencies
  - checks_remote_test.go: rename test, drop the pinned check
    (7 -> 6 remote checks today)
  - checks_connections.go / checks_toolboxes.go: drop stale
    cross-reference comments to the removed check
  - doctor_format.go: drop `remote.model-deployments` from the
    `azd provision` remediation case
  - doctor_format_test.go: swap fixture to `remote.connections` so
    the dedup test still has two checks -> azd provision

Added:
  - internal/cmd/doctor/shared_test.go: host fixedAssembler,
    fixedProjectIDReader, and validProjectResourceID -- these helpers
    used to be co-located with the deleted test file but are also
    consumed by checks_connections_test.go and checks_toolboxes_test.go.

Kept (state.ModelRefs / state.HasModels and the manifest walker):
unconsumed in production but cheap, and the redesign will likely
reuse them.
