+++ cli/azd/extensions/azure.ai.agents/internal/cmd/doctor/checks_remote_test.go
@@ -12,35 +12,38 @@ import (
 
 // ---- NewRemoteChecks contract ----
 
-// TestNewRemoteChecks_HasAuthFoundryEndpointRBACAgentStatusIdentityRolesModelDeploymentsConnections
-// pins the current shape of the remote chain: exactly seven checks,
+// TestNewRemoteChecks_HasAuthFoundryEndpointRBACAgentStatusIdentityRolesConnections
+// pins the current shape of the remote chain: exactly six checks,
 // in the order `remote.auth` → `remote.foundry-endpoint` →
 // `remote.rbac` → `remote.agent-status` → `remote.agent-identity-roles`
-// → `remote.model-deployments` → `remote.connections`, all with
-// Remote=true. The ordering matters because `remote.foundry-endpoint`
-// skip-cascades against `remote.auth`'s prior Result, `remote.rbac`
-// skip-cascades against `remote.auth` (but NOT
-// `remote.foundry-endpoint`, per the design's dependency matrix line
-// 115 — RBAC reads ARM, not the data plane), `remote.agent-status`
-// skip-cascades against `remote.auth` + `remote.foundry-endpoint`
-// (Reader-level Foundry call, deliberately bypasses RBAC),
-// `remote.agent-identity-roles` cascades against `remote.agent-status`
-// Pass so the per-agent role enumeration only runs against agents the
-// previous check confirmed active, `remote.model-deployments`
-// cascades against `remote.auth` + `remote.foundry-endpoint` because
-// it issues an ARM-side Cognitive Services deployments list using the
-// same identity, and `remote.connections` cascades against the same
-// pair plus `state.HasConnections` because it lists Foundry
-// connections via the data plane. Any future re-ordering or
-// insertion has to come through this assertion.
-func TestNewRemoteChecks_HasAuthFoundryEndpointRBACAgentStatusIdentityRolesModelDeploymentsConnections(t *testing.T) {
+// → `remote.connections`, all with Remote=true. The ordering matters
+// because `remote.foundry-endpoint` skip-cascades against
+// `remote.auth`'s prior Result, `remote.rbac` skip-cascades against
+// `remote.auth` (but NOT `remote.foundry-endpoint`, per the design's
+// dependency matrix — RBAC reads ARM, not the data plane),
+// `remote.agent-status` skip-cascades against `remote.auth` +
+// `remote.foundry-endpoint` (Reader-level Foundry call, deliberately
+// bypasses RBAC), `remote.agent-identity-roles` cascades against
+// `remote.agent-status` Pass so the per-agent role enumeration only
+// runs against agents the previous check confirmed active, and
+// `remote.connections` cascades against `remote.auth` +
+// `remote.foundry-endpoint` plus `state.HasConnections` because it
+// lists Foundry connections via the data plane. Any future
+// re-ordering or insertion has to come through this assertion.
+//
+// Note: a `remote.model-deployments` check used to sit between
+// agent-identity-roles and connections; it was removed pending a
+// redesign because its comparison was incorrect — the manifest's
+// `resources[].name` is a logical alias, not a Foundry deployment
+// name.
+func TestNewRemoteChecks_HasAuthFoundryEndpointRBACAgentStatusIdentityRolesConnections(t *testing.T) {
 	t.Parallel()
 
 	got := NewRemoteChecks(Dependencies{})
 
-	require.Len(t, got, 7,
+	require.Len(t, got, 6,
 		"NewRemoteChecks should contain auth, foundry-endpoint, rbac, agent-status, "+
-			"agent-identity-roles, model-deployments, and connections today")
+			"agent-identity-roles, and connections today")
 	require.Equal(t, "remote.auth", got[0].ID)
 	require.Equal(t, "authentication", got[0].Name)
 	require.True(t, got[0].Remote, "remote.auth must declare Remote=true")
@@ -61,14 +64,10 @@ func TestNewRemoteChecks_HasAuthFoundryEndpointRBACAgentStatusIdentityRolesModel
 	require.Equal(t, "Agent identity role assignments", got[4].Name)
 	require.True(t, got[4].Remote, "remote.agent-identity-roles must declare Remote=true")
 	require.NotNil(t, got[4].Fn, "remote.agent-identity-roles must have a non-nil Fn")
-	require.Equal(t, "remote.model-deployments", got[5].ID)
-	require.Equal(t, "Manifest model deployments exist in Foundry", got[5].Name)
-	require.True(t, got[5].Remote, "remote.model-deployments must declare Remote=true")
-	require.NotNil(t, got[5].Fn, "remote.model-deployments must have a non-nil Fn")
-	require.Equal(t, "remote.connections", got[6].ID)
-	require.Equal(t, "Manifest connections exist on Foundry project", got[6].Name)
-	require.True(t, got[6].Remote, "remote.connections must declare Remote=true")
-	require.NotNil(t, got[6].Fn, "remote.connections must have a non-nil Fn")
+	require.Equal(t, "remote.connections", got[5].ID)
+	require.Equal(t, "Manifest connections exist on Foundry project", got[5].Name)
+	require.True(t, got[5].Remote, "remote.connections must declare Remote=true")
+	require.NotNil(t, got[5].Fn, "remote.connections must have a non-nil Fn")
 }
 
 // TestNewLocalAndRemoteChecks_ProductionCompositionLocalsFirst pins the
