+++ cli/azd/extensions/azure.ai.agents/internal/cmd/doctor_format_test.go
@@ -234,8 +234,8 @@ func TestPrintDoctorReportText_ToFixBlockOnFailure(t *testing.T) {
 			{ID: "remote.auth", Name: "authentication", Status: doctor.StatusPass},
 			{ID: "remote.foundry-endpoint", Name: "Foundry endpoint",
 				Status: doctor.StatusFail, Message: "endpoint unreachable"},
-			{ID: "remote.model-deployments", Name: "model deployments",
-				Status: doctor.StatusFail, Message: "model missing"},
+			{ID: "remote.connections", Name: "connections",
+				Status: doctor.StatusFail, Message: "connection missing"},
 			{ID: "remote.agent-status", Name: "agents active",
 				Status: doctor.StatusFail, Message: "1 of 1 agents have not been deployed"},
 		},
